common: move troika and ftroika to crypto
