+++ common/crypto/BUILD
@@ -1,0 +1,1 @@
+package(default_visibility = ["//visibility:public"])
