+++ common/crypto/ftroika/general.h
